Refactor existing validation to make future validation possible
