+++ lib/maze/schemas/error_validator.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require_relative '../helper'
+
+module Maze
+  module Schemas
+
+    # Contains a set of pre-defined validations for ensuring errors are correct
+    class ErrorValidator
+
+      # Runs the validation against the trace given
+      def validate
+        @success = true
+      end
+    end
+  end
+end
